Allow common annotations to use default max values
Add configuration for setting the default max results values.
Allow truthy values to #annotate to use the default max result values.
