+++ test/gcloud/vision/project_test.rb
@@ -33,655 +33,6 @@ describe Gcloud::Vision::Project, :mock_vision do
     image.wont_be :url?
   end
 
-  it "detects face detection" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      face = requests.first
-      face["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      face["features"].count.must_equal 1
-      face["features"].first["type"].must_equal "FACE_DETECTION"
-      face["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       face_response_json]
-    end
-
-    annotation = vision.annotate filepath, faces: 1
-    annotation.wont_be :nil?
-    annotation.face.wont_be :nil?
-  end
-
-  it "detects face detection using mark alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      face = requests.first
-      face["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      face["features"].count.must_equal 1
-      face["features"].first["type"].must_equal "FACE_DETECTION"
-      face["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       face_response_json]
-    end
-
-    annotation = vision.mark filepath, faces: 1
-    annotation.wont_be :nil?
-    annotation.face.wont_be :nil?
-  end
-
-  it "detects face detection using detect alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      face = requests.first
-      face["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      face["features"].count.must_equal 1
-      face["features"].first["type"].must_equal "FACE_DETECTION"
-      face["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       face_response_json]
-    end
-
-    annotation = vision.detect filepath, faces: 1
-    annotation.wont_be :nil?
-    annotation.face.wont_be :nil?
-  end
-
-  it "detects face detection on multiple images" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 2
-      requests.first["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.first["features"].count.must_equal 1
-      requests.first["features"].first["type"].must_equal "FACE_DETECTION"
-      requests.first["features"].first["maxResults"].must_equal 1
-      requests.last["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.last["features"].count.must_equal 1
-      requests.last["features"].first["type"].must_equal "FACE_DETECTION"
-      requests.last["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       faces_response_json]
-    end
-
-    annotations = vision.annotate filepath, filepath, faces: 1
-    annotations.count.must_equal 2
-    annotations.first.face.wont_be :nil?
-    annotations.last.face.wont_be :nil?
-  end
-
-  it "detects landmark detection" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      landmark = requests.first
-      landmark["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      landmark["features"].count.must_equal 1
-      landmark["features"].first["type"].must_equal "LANDMARK_DETECTION"
-      landmark["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       landmark_response_json]
-    end
-
-    annotation = vision.annotate filepath, landmarks: 1
-    annotation.wont_be :nil?
-    annotation.landmark.wont_be :nil?
-  end
-
-  it "detects landmark detection using mark alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      landmark = requests.first
-      landmark["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      landmark["features"].count.must_equal 1
-      landmark["features"].first["type"].must_equal "LANDMARK_DETECTION"
-      landmark["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       landmark_response_json]
-    end
-
-    annotation = vision.mark filepath, landmarks: 1
-    annotation.wont_be :nil?
-    annotation.landmark.wont_be :nil?
-  end
-
-  it "detects landmark detection using detect alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      landmark = requests.first
-      landmark["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      landmark["features"].count.must_equal 1
-      landmark["features"].first["type"].must_equal "LANDMARK_DETECTION"
-      landmark["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       landmark_response_json]
-    end
-
-    annotation = vision.detect filepath, landmarks: 1
-    annotation.wont_be :nil?
-    annotation.landmark.wont_be :nil?
-  end
-
-  it "detects landmark detection on multiple images" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 2
-      requests.first["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.first["features"].count.must_equal 1
-      requests.first["features"].first["type"].must_equal "LANDMARK_DETECTION"
-      requests.first["features"].first["maxResults"].must_equal 1
-      requests.last["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.last["features"].count.must_equal 1
-      requests.last["features"].first["type"].must_equal "LANDMARK_DETECTION"
-      requests.last["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       landmarks_response_json]
-    end
-
-    annotations = vision.annotate filepath, filepath, landmarks: 1
-    annotations.count.must_equal 2
-    annotations.first.landmark.wont_be :nil?
-    annotations.last.landmark.wont_be :nil?
-  end
-
-  it "detects logo detection" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      logo = requests.first
-      logo["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      logo["features"].count.must_equal 1
-      logo["features"].first["type"].must_equal "LOGO_DETECTION"
-      logo["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       logo_response_json]
-    end
-
-    annotation = vision.annotate filepath, logos: 1
-    annotation.wont_be :nil?
-    annotation.logo.wont_be :nil?
-  end
-
-  it "detects logo detection using mark alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      logo = requests.first
-      logo["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      logo["features"].count.must_equal 1
-      logo["features"].first["type"].must_equal "LOGO_DETECTION"
-      logo["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       logo_response_json]
-    end
-
-    annotation = vision.mark filepath, logos: 1
-    annotation.wont_be :nil?
-    annotation.logo.wont_be :nil?
-  end
-
-  it "detects logo detection using detect alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      logo = requests.first
-      logo["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      logo["features"].count.must_equal 1
-      logo["features"].first["type"].must_equal "LOGO_DETECTION"
-      logo["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       logo_response_json]
-    end
-
-    annotation = vision.detect filepath, logos: 1
-    annotation.wont_be :nil?
-    annotation.logo.wont_be :nil?
-  end
-
-  it "detects logo detection on multiple images" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 2
-      requests.first["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.first["features"].count.must_equal 1
-      requests.first["features"].first["type"].must_equal "LOGO_DETECTION"
-      requests.first["features"].first["maxResults"].must_equal 1
-      requests.last["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.last["features"].count.must_equal 1
-      requests.last["features"].first["type"].must_equal "LOGO_DETECTION"
-      requests.last["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       logos_response_json]
-    end
-
-    annotations = vision.annotate filepath, filepath, logos: 1
-    annotations.count.must_equal 2
-    annotations.first.logo.wont_be :nil?
-    annotations.last.logo.wont_be :nil?
-  end
-
-  it "detects label detection" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      label = requests.first
-      label["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      label["features"].count.must_equal 1
-      label["features"].first["type"].must_equal "LABEL_DETECTION"
-      label["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       label_response_json]
-    end
-
-    annotation = vision.annotate filepath, labels: 1
-    annotation.wont_be :nil?
-    annotation.label.wont_be :nil?
-  end
-
-  it "detects label detection using mark alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      label = requests.first
-      label["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      label["features"].count.must_equal 1
-      label["features"].first["type"].must_equal "LABEL_DETECTION"
-      label["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       label_response_json]
-    end
-
-    annotation = vision.mark filepath, labels: 1
-    annotation.wont_be :nil?
-    annotation.label.wont_be :nil?
-  end
-
-  it "detects label detection using detect alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      label = requests.first
-      label["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      label["features"].count.must_equal 1
-      label["features"].first["type"].must_equal "LABEL_DETECTION"
-      label["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       label_response_json]
-    end
-
-    annotation = vision.detect filepath, labels: 1
-    annotation.wont_be :nil?
-    annotation.label.wont_be :nil?
-  end
-
-  it "detects label detection on multiple images" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 2
-      requests.first["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.first["features"].count.must_equal 1
-      requests.first["features"].first["type"].must_equal "LABEL_DETECTION"
-      requests.first["features"].first["maxResults"].must_equal 1
-      requests.last["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.last["features"].count.must_equal 1
-      requests.last["features"].first["type"].must_equal "LABEL_DETECTION"
-      requests.last["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       labels_response_json]
-    end
-
-    annotations = vision.annotate filepath, filepath, labels: 1
-    annotations.count.must_equal 2
-    annotations.first.label.wont_be :nil?
-    annotations.last.label.wont_be :nil?
-  end
-
-  it "detects text detection" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      text = requests.first
-      text["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      text["features"].count.must_equal 1
-      text["features"].first["type"].must_equal "TEXT_DETECTION"
-      text["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       text_response_json]
-    end
-
-    annotation = vision.annotate filepath, text: true
-    annotation.wont_be :nil?
-    annotation.text.wont_be :nil?
-    annotation.text.text.must_include "Google Cloud Client Library for Ruby"
-    annotation.text.locale.must_equal "en"
-    annotation.text.words.count.must_equal 28
-    annotation.text.words[0].text.must_equal "Google"
-    annotation.text.words[0].bounds.map(&:to_a).must_equal [[13, 8], [53, 8], [53, 23], [13, 23]]
-    annotation.text.words[27].text.must_equal "Storage."
-    annotation.text.words[27].bounds.map(&:to_a).must_equal [[304, 59], [351, 59], [351, 74], [304, 74]]
-  end
-
-  it "detects text detection using mark alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      text = requests.first
-      text["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      text["features"].count.must_equal 1
-      text["features"].first["type"].must_equal "TEXT_DETECTION"
-      text["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       text_response_json]
-    end
-
-    annotation = vision.mark filepath, text: true
-    annotation.wont_be :nil?
-    annotation.text.wont_be :nil?
-  end
-
-  it "detects text detection using detect alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      text = requests.first
-      text["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      text["features"].count.must_equal 1
-      text["features"].first["type"].must_equal "TEXT_DETECTION"
-      text["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       text_response_json]
-    end
-
-    annotation = vision.detect filepath, text: true
-    annotation.wont_be :nil?
-    annotation.text.wont_be :nil?
-  end
-
-  it "detects text detection on multiple images" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 2
-      requests.first["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.first["features"].count.must_equal 1
-      requests.first["features"].first["type"].must_equal "TEXT_DETECTION"
-      requests.first["features"].first["maxResults"].must_equal 1
-      requests.last["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.last["features"].count.must_equal 1
-      requests.last["features"].first["type"].must_equal "TEXT_DETECTION"
-      requests.last["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       texts_response_json]
-    end
-
-    annotations = vision.annotate filepath, filepath, text: true
-    annotations.count.must_equal 2
-    annotations.first.text.wont_be :nil?
-    annotations.last.text.wont_be :nil?
-  end
-
-  it "detects safe_search detection" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      safe_search = requests.first
-      safe_search["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      safe_search["features"].count.must_equal 1
-      safe_search["features"].first["type"].must_equal "SAFE_SEARCH_DETECTION"
-      safe_search["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       safe_search_response_json]
-    end
-
-    annotation = vision.annotate filepath, safe_search: true
-    annotation.wont_be :nil?
-
-    annotation.safe_search.wont_be :nil?
-    annotation.safe_search.wont_be :adult?
-    annotation.safe_search.wont_be :spoof?
-    annotation.safe_search.must_be :medical?
-    annotation.safe_search.must_be :violence?
-  end
-
-  it "detects safe_search detection using mark alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      safe_search = requests.first
-      safe_search["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      safe_search["features"].count.must_equal 1
-      safe_search["features"].first["type"].must_equal "SAFE_SEARCH_DETECTION"
-      safe_search["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       safe_search_response_json]
-    end
-
-    annotation = vision.mark filepath, safe_search: true
-    annotation.wont_be :nil?
-
-    annotation.safe_search.wont_be :nil?
-    annotation.safe_search.wont_be :adult?
-    annotation.safe_search.wont_be :spoof?
-    annotation.safe_search.must_be :medical?
-    annotation.safe_search.must_be :violence?
-  end
-
-  it "detects safe_search detection using detect alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      safe_search = requests.first
-      safe_search["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      safe_search["features"].count.must_equal 1
-      safe_search["features"].first["type"].must_equal "SAFE_SEARCH_DETECTION"
-      safe_search["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       safe_search_response_json]
-    end
-
-    annotation = vision.detect filepath, safe_search: true
-    annotation.wont_be :nil?
-
-    annotation.safe_search.wont_be :nil?
-    annotation.safe_search.wont_be :adult?
-    annotation.safe_search.wont_be :spoof?
-    annotation.safe_search.must_be :medical?
-    annotation.safe_search.must_be :violence?
-  end
-
-  it "detects safe_search detection on multiple images" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 2
-      requests.first["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.first["features"].count.must_equal 1
-      requests.first["features"].first["type"].must_equal "SAFE_SEARCH_DETECTION"
-      requests.first["features"].first["maxResults"].must_equal 1
-      requests.last["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.last["features"].count.must_equal 1
-      requests.last["features"].first["type"].must_equal "SAFE_SEARCH_DETECTION"
-      requests.last["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       safe_searchs_response_json]
-    end
-
-    annotations = vision.annotate filepath, filepath, safe_search: true
-    annotations.count.must_equal 2
-
-    annotations.first.safe_search.wont_be :nil?
-    annotations.first.safe_search.wont_be :adult?
-    annotations.first.safe_search.wont_be :spoof?
-    annotations.first.safe_search.must_be :medical?
-    annotations.first.safe_search.must_be :violence?
-
-    annotations.last.safe_search.wont_be :nil?
-    annotations.last.safe_search.wont_be :adult?
-    annotations.last.safe_search.wont_be :spoof?
-    annotations.last.safe_search.must_be :medical?
-    annotations.last.safe_search.must_be :violence?
-  end
-
-  it "detects properties detection" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      properties = requests.first
-      properties["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      properties["features"].count.must_equal 1
-      properties["features"].first["type"].must_equal "IMAGE_PROPERTIES"
-      properties["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       properties_response_json]
-    end
-
-    annotation = vision.annotate filepath, properties: true
-    annotation.wont_be :nil?
-
-    annotation.properties.colors.count.must_equal 10
-
-    annotation.properties.colors[0].red.must_equal 145
-    annotation.properties.colors[0].green.must_equal 193
-    annotation.properties.colors[0].blue.must_equal 254
-    annotation.properties.colors[0].alpha.must_equal 1.0
-    annotation.properties.colors[0].rgb.must_equal "91c1fe"
-    annotation.properties.colors[0].score.must_equal 0.65757853
-    annotation.properties.colors[0].pixel_fraction.must_equal 0.16903226
-
-    annotation.properties.colors[9].red.must_equal 156
-    annotation.properties.colors[9].green.must_equal 214
-    annotation.properties.colors[9].blue.must_equal 255
-    annotation.properties.colors[9].alpha.must_equal 1.0
-    annotation.properties.colors[9].rgb.must_equal "9cd6ff"
-    annotation.properties.colors[9].score.must_equal 0.00096750073
-    annotation.properties.colors[9].pixel_fraction.must_equal 0.00064516132
-  end
-
-  it "detects properties detection using mark alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      properties = requests.first
-      properties["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      properties["features"].count.must_equal 1
-      properties["features"].first["type"].must_equal "IMAGE_PROPERTIES"
-      properties["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       properties_response_json]
-    end
-
-    annotation = vision.mark filepath, properties: true
-    annotation.wont_be :nil?
-
-    annotation.properties.colors.count.must_equal 10
-
-    annotation.properties.colors[0].red.must_equal 145
-    annotation.properties.colors[0].green.must_equal 193
-    annotation.properties.colors[0].blue.must_equal 254
-    annotation.properties.colors[0].alpha.must_equal 1.0
-    annotation.properties.colors[0].rgb.must_equal "91c1fe"
-    annotation.properties.colors[0].score.must_equal 0.65757853
-    annotation.properties.colors[0].pixel_fraction.must_equal 0.16903226
-
-    annotation.properties.colors[9].red.must_equal 156
-    annotation.properties.colors[9].green.must_equal 214
-    annotation.properties.colors[9].blue.must_equal 255
-    annotation.properties.colors[9].alpha.must_equal 1.0
-    annotation.properties.colors[9].rgb.must_equal "9cd6ff"
-    annotation.properties.colors[9].score.must_equal 0.00096750073
-    annotation.properties.colors[9].pixel_fraction.must_equal 0.00064516132
-  end
-
-  it "detects properties detection using detect alias" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 1
-      properties = requests.first
-      properties["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      properties["features"].count.must_equal 1
-      properties["features"].first["type"].must_equal "IMAGE_PROPERTIES"
-      properties["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       properties_response_json]
-    end
-
-    annotation = vision.detect filepath, properties: true
-    annotation.wont_be :nil?
-
-    annotation.properties.colors.count.must_equal 10
-
-    annotation.properties.colors[0].red.must_equal 145
-    annotation.properties.colors[0].green.must_equal 193
-    annotation.properties.colors[0].blue.must_equal 254
-    annotation.properties.colors[0].alpha.must_equal 1.0
-    annotation.properties.colors[0].rgb.must_equal "91c1fe"
-    annotation.properties.colors[0].score.must_equal 0.65757853
-    annotation.properties.colors[0].pixel_fraction.must_equal 0.16903226
-
-    annotation.properties.colors[9].red.must_equal 156
-    annotation.properties.colors[9].green.must_equal 214
-    annotation.properties.colors[9].blue.must_equal 255
-    annotation.properties.colors[9].alpha.must_equal 1.0
-    annotation.properties.colors[9].rgb.must_equal "9cd6ff"
-    annotation.properties.colors[9].score.must_equal 0.00096750073
-    annotation.properties.colors[9].pixel_fraction.must_equal 0.00064516132
-  end
-
-  it "detects properties detection on multiple images" do
-    mock_connection.post "/v1/images:annotate" do |env|
-      requests = JSON.parse(env.body)["requests"]
-      requests.count.must_equal 2
-      requests.first["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.first["features"].count.must_equal 1
-      requests.first["features"].first["type"].must_equal "IMAGE_PROPERTIES"
-      requests.first["features"].first["maxResults"].must_equal 1
-      requests.last["image"]["content"].must_equal Base64.encode64(File.read(filepath, mode: "rb"))
-      requests.last["features"].count.must_equal 1
-      requests.last["features"].first["type"].must_equal "IMAGE_PROPERTIES"
-      requests.last["features"].first["maxResults"].must_equal 1
-      [200, {"Content-Type" => "application/json"},
-       propertiess_response_json]
-    end
-
-    annotations = vision.annotate filepath, filepath, properties: true
-    annotations.count.must_equal 2
-
-    annotations[0].properties.colors.count.must_equal 10
-
-    annotations[0].properties.colors[0].red.must_equal 145
-    annotations[0].properties.colors[0].green.must_equal 193
-    annotations[0].properties.colors[0].blue.must_equal 254
-    annotations[0].properties.colors[0].alpha.must_equal 1.0
-    annotations[0].properties.colors[0].rgb.must_equal "91c1fe"
-    annotations[0].properties.colors[0].score.must_equal 0.65757853
-    annotations[0].properties.colors[0].pixel_fraction.must_equal 0.16903226
-
-    annotations[0].properties.colors[9].red.must_equal 156
-    annotations[0].properties.colors[9].green.must_equal 214
-    annotations[0].properties.colors[9].blue.must_equal 255
-    annotations[0].properties.colors[9].alpha.must_equal 1.0
-    annotations[0].properties.colors[9].rgb.must_equal "9cd6ff"
-    annotations[0].properties.colors[9].score.must_equal 0.00096750073
-    annotations[0].properties.colors[9].pixel_fraction.must_equal 0.00064516132
-
-    annotations[1].properties.colors.count.must_equal 10
-
-    annotations[1].properties.colors[0].red.must_equal 145
-    annotations[1].properties.colors[0].green.must_equal 193
-    annotations[1].properties.colors[0].blue.must_equal 254
-    annotations[1].properties.colors[0].alpha.must_equal 1.0
-    annotations[1].properties.colors[0].rgb.must_equal "91c1fe"
-    annotations[1].properties.colors[0].score.must_equal 0.65757853
-    annotations[1].properties.colors[0].pixel_fraction.must_equal 0.16903226
-
-    annotations[1].properties.colors[9].red.must_equal 156
-    annotations[1].properties.colors[9].green.must_equal 214
-    annotations[1].properties.colors[9].blue.must_equal 255
-    annotations[1].properties.colors[9].alpha.must_equal 1.0
-    annotations[1].properties.colors[9].rgb.must_equal "9cd6ff"
-    annotations[1].properties.colors[9].score.must_equal 0.00096750073
-    annotations[1].properties.colors[9].pixel_fraction.must_equal 0.00064516132
-  end
-
   it "allows different annotation options for different images" do
     mock_connection.post "/v1/images:annotate" do |env|
       requests = JSON.parse(env.body)["requests"]
@@ -936,128 +287,12 @@ describe Gcloud::Vision::Project, :mock_vision do
     end
   end
 
-  def face_response_json
-    {
-      responses: [{
-        faceAnnotations: [face_annotation_response]
-      }]
-    }.to_json
-  end
-
   def faces_response_json
     {
       responses: [{
         faceAnnotations: [face_annotation_response]
       }, {
         faceAnnotations: [face_annotation_response]
-      }]
-    }.to_json
-  end
-
-  def landmark_response_json
-    {
-      responses: [{
-        landmarkAnnotations: [landmark_annotation_response]
-      }]
-    }.to_json
-  end
-
-  def landmarks_response_json
-    {
-      responses: [{
-        landmarkAnnotations: [landmark_annotation_response]
-      }, {
-        landmarkAnnotations: [landmark_annotation_response]
-      }]
-    }.to_json
-  end
-
-  def logo_response_json
-    {
-      responses: [{
-        logoAnnotations: [logo_annotation_response]
-      }]
-    }.to_json
-  end
-
-  def logos_response_json
-    {
-      responses: [{
-        logoAnnotations: [logo_annotation_response]
-      }, {
-        logoAnnotations: [logo_annotation_response]
-      }]
-    }.to_json
-  end
-
-  def label_response_json
-    {
-      responses: [{
-        labelAnnotations: [label_annotation_response]
-      }]
-    }.to_json
-  end
-
-  def labels_response_json
-    {
-      responses: [{
-        labelAnnotations: [label_annotation_response]
-      }, {
-        labelAnnotations: [label_annotation_response]
-      }]
-    }.to_json
-  end
-
-  def text_response_json
-    {
-      responses: [{
-        textAnnotations: text_annotation_responses
-      }]
-    }.to_json
-  end
-
-  def texts_response_json
-    {
-      responses: [{
-        textAnnotations: text_annotation_responses
-      }, {
-        textAnnotations: text_annotation_responses
-      }]
-    }.to_json
-  end
-
-  def safe_search_response_json
-    {
-      responses: [{
-        safeSearchAnnotation: safe_search_annotation_response
-      }]
-    }.to_json
-  end
-
-  def safe_searchs_response_json
-    {
-      responses: [{
-        safeSearchAnnotation: safe_search_annotation_response
-      }, {
-        safeSearchAnnotation: safe_search_annotation_response
-      }]
-    }.to_json
-  end
-
-  def properties_response_json
-    {
-      responses: [{
-        imagePropertiesAnnotation: properties_annotation_response
-      }]
-    }.to_json
-  end
-
-  def propertiess_response_json
-    {
-      responses: [{
-        imagePropertiesAnnotation: properties_annotation_response
-      }, {
-        imagePropertiesAnnotation: properties_annotation_response
       }]
     }.to_json
   end
